% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayolaTAMM-package.R
\docType{package}
\name{crayolaTAMM-package}
\alias{crayolaTAMM}
\alias{crayolaTAMM-package}
\title{crayolaTAMM: Create "Crayola" Summary of TAMM File}
\description{
For a Chinook TAMM file from the FRAM modeling process, produces a file summarizing the impacts for each key stock, as well as identifying the fisheries that most impact that stock.
}
\author{
\strong{Maintainer}: Robert Roose \email{rroose@stillaguamish.com} (\href{https://orcid.org/0000-0003-3785-5224}{ORCID})

Authors:
\itemize{
  \item Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})
}

}
\keyword{internal}
