% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-crayola-tamm.R
\name{make_crayola_tamm}
\alias{make_crayola_tamm}
\title{Make workbook summary of Chinook TAMM}
\usage{
make_crayola_tamm(
  filepath,
  output_path = NULL,
  plot_width = 10,
  plot_height = 8,
  plot_units = "in",
  verbose = TRUE
)
}
\arguments{
\item{filepath}{Filepath of TAMM, including \code{.xlsx} suffix}

\item{output_path}{Filepath for resulting file. Optional; if not provided, will save in same folder as \code{filepath} with
name of "\{filepath\}_Impacts_crayola.xlsx"}

\item{plot_width}{Controls width of barplots in stock sheets. Numeric, defaults to 10}

\item{plot_height}{Controls height of barplots in stock sheets. Numeric, defaults to 8}

\item{plot_units}{Character string defining units of \code{plot_width} and \code{plot_height}. Defaults to "in", can also be "cm" or "px".}

\item{verbose}{Should progress messages and saved filepath be printed to screen?}
}
\value{
Invisibly returns filepath of the summary file.
}
\description{
Make workbook summary of Chinook TAMM
}
